move sdk.init to test/cli_tests
